langchain-rag-pineCone/pyproject.toml
[project]
name = "langchain-rag-pinecone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.1.4",
    "langchain-pinecone>=0.2.6",
    "python-dotenv>=1.1.0",
]
